refactor the test photos
