+++ app/src/androidTest/java/com/epfl/drawyourpath/database/MockDatabaseTestPhoto.kt
@@ -3,9 +3,13 @@ package com.epfl.drawyourpath.database
 import android.graphics.Bitmap
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.epfl.Utils.drawyourpath.Utils
+import com.epfl.drawyourpath.chat.Message
 import org.junit.Assert
 import org.junit.Test
 import org.junit.runner.RunWith
+import java.time.LocalDate
+import java.time.LocalTime
+import java.time.ZoneOffset
 
 @RunWith(AndroidJUnit4::class)
 class MockDatabaseTestPhoto {
@@ -25,6 +29,29 @@ class MockDatabaseTestPhoto {
             Utils.encodePhoto(
                 Utils.decodePhoto(database.users[userIdTest]?.picture!!),
             ),
+        )
+    }
+
+    /**
+     * Test that a picture message is correctly added to a conversation
+     */
+    @Test
+    fun addPictureMessageCorrectly(){
+        val database = MockDatabase()
+        val conversationId = database.MOCK_CHAT_MESSAGES[0].conversationId!!
+        val senderId = database.MOCK_USERS[0].userId!!
+        val date = LocalDate.now().atTime(LocalTime.now()).toEpochSecond(ZoneOffset.UTC)
+        val messageSent = Message.createPictureMessage(senderId, photoProfile, date)
+        database.addChatMessage(conversationId, messageSent)
+        //check the chat messages list
+        Assert.assertEquals(
+            listOf(messageSent) + (database.MOCK_CHAT_MESSAGES[0].chat ?: emptyList()),
+            database.chatMessages[conversationId]!!.chat
+        )
+        //check the chat preview
+        Assert.assertEquals(
+            database.MOCK_CHAT_PREVIEWS[0].copy(lastMessage = messageSent),
+            database.chatPreviews[conversationId]
         )
     }
 }
